Passando do lcc2 para nuvem

+++ src/main/Tutoria.java
@@ -13,7 +13,7 @@ import java.util.Map;
 public class Tutoria {
 
 	private int dinheiroTutor;
-	private int pontuacaoAtual;
+	private double pontuacaoAtual;
 	private int quantia;
 	private Map<String, Disciplina> disciplinas;
 	private Disponibilidade disponibilidade;
@@ -91,7 +91,7 @@ public class Tutoria {
 	 * 
 	 * @return int referente a avaliação da tutoria.
 	 */
-	public int getAvaliacao() {
+	public double getAvaliacao() {
 		return pontuacaoAtual;
 	}
 
@@ -183,7 +183,7 @@ public class Tutoria {
 	public int hashCode() {
 		final int prime = 31;
 		int result = 1;
-		result = prime * result + pontuacaoAtual;
+		result = (int)(prime * result + pontuacaoAtual) ;
 		result = prime * result + ((disciplinas == null) ? 0 : disciplinas.hashCode());
 		result = prime * result + quantia;
 		return result;
